+++ guillotina_cms/__init__.py
@@ -8,11 +8,14 @@ from guillotina.i18n import MessageFactory
 
 _ = MessageFactory("guillotina_cms")
 
-
 app_settings = {
-    "applications": ["guillotina.contrib.catalog.pg", "guillotina.contrib.swagger", "guillotina_dbusers"],
-    "available_tiles": {},
-    "commands": {"upgrade": "guillotina_cms.commands.upgrade.UpgradeCommand"},
+    "applications": [
+        "guillotina.contrib.catalog.pg",
+        "guillotina.contrib.swagger",
+        "guillotina.contrib.dbusers",
+    ],
+    "available_blocks": {},
+    "commands": {"create-container": "guillotina_cms.commands.create.CMSCreateCommand"},
     "load_utilities": {
         "guillotina_cms.workflow": {
             "provides": "guillotina_cms.interfaces.IWorkflowUtility",
@@ -31,16 +34,19 @@ app_settings = {
             "event_listing",
             "document_view",
         ],
-        "Document": ["document_view", "layout_view"],
+        "Document": ["document_view", "layout_view", "default"],
         "Container": ["document_view", "layout_view"],
         "News": ["document_view", "layout_view"],
         "Event": ["document_view", "layout_view"],
         "Link": ["document_view", "layout_view"],
         "File": ["document_view", "layout_view"],
-        "Image": ["document_view", "layout_view"]
+        "Image": ["document_view", "layout_view"],
     },
     "workflows": {
-        "private": {"initial_state": "private", "states": {"private": {"set_permission": {}, "actions": {}}}},
+        "private": {
+            "initial_state": "private",
+            "states": {"private": {"set_permission": {}, "actions": {}}},
+        },
         "basic": {
             "initial_state": "private",
             "states": {
@@ -100,14 +106,14 @@ app_settings = {
         "guillotina_cms.content.image.IImage": "basic",
         "guillotina_cms.content.folder.IFolder": "basic",
     },
-    "default_tiles": {
+    "default_blocks": {
         "Document": {
-            "tiles": {"tile1": {"@type": "title"}, "tile2": {"@type": "text"}},
-            "tiles_layout": {"items": ["tile1", "tile2"]},
+            "blocks": {"tile1": {"@type": "title"}, "tile2": {"@type": "text"}},
+            "blocks_layout": {"items": ["tile1", "tile2"]},
         },
         "Container": {
-            "tiles": {"tile1": {"@type": "title"}, "tile2": {"@type": "text"}},
-            "tiles_layout": {"items": ["tile1", "tile2"]},
+            "blocks": {"tile1": {"@type": "title"}, "tile2": {"@type": "text"}},
+            "blocks_layout": {"items": ["tile1", "tile2"]},
         },
     },
     "global_disallowed_types": [
@@ -145,7 +151,7 @@ def includeme(root, settings):
     configure.scan("guillotina_cms.permissions")
     configure.scan("guillotina_cms.install")
     configure.scan("guillotina_cms.subscribers")
-    configure.scan("guillotina_cms.tiles")
+    configure.scan("guillotina_cms.blocks")
 
     if "guillotina_elasticsearch" in settings.get("applications", []):
         if "load_utilities" not in settings: