+++ guillotina_cms/content/__init__.py
@@ -1,6 +1,6 @@
 from .document import Document  # noqa
 from .link import Link  # noqa
-from .news import News  # noqa
+from .event import Event  # noqa
 from .image import Image  # noqa
 from .file import File  # noqa
 from .folder import CMSFolder  # noqa