+++ guillotina_cms/content/document.py
@@ -11,19 +11,10 @@ from guillotina_cms.interfaces import IDocument
     behaviors=[
         "guillotina.behaviors.dublincore.IDublinCore",
         "guillotina_cms.interfaces.base.ICMSBehavior",
-        "guillotina_cms.interfaces.tiles.ITiles",
+        "guillotina_cms.interfaces.blocks.IBlocks",
     ],
     allowed_types=["Image", "File"],  # dynamically calculated
 )
 class Document(Folder):
     pass
 
-
-@index.with_accessor(IDocument, "text", type="searchabletext")
-def get_text_from_richtext(ob):
-    # Richtext is a dict and we only care about the text
-    try:
-        if ob.text is not None:
-            return ob.text.data
-    except AttributeError:
-        pass