+++ guillotina_cms/fields/__init__.py
@@ -1,1 +1,1 @@
-from . import richtext  # noqa
+from .richtext import RichTextField  # noqa