+++ CHANGELOG.rst
@@ -5,7 +5,8 @@
 - Removing @order endpoint right now for not beeing compatible G5
 - Images and folder should be workflow enabled
 - Choice converter fix
-
+- Rename tiles -> blocks
+  [sneridagh]
 
 5.0.0a1 (2019-06-22)
 --------------------